src/predict/pyproject.toml
[tool.poetry]
name = "predict"
version = "0.1.0"
description = ""
authors = ["Duncan Haywood"]
license = "Prop"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.3.0"
python-json-logger = "^2.0.4"
sagemaker = "^2.126.0"
boto3 = "^1.26.36"
pytest = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
